abstract the common adapter code from the postgres adapter
